redux化 components/chatbot.jsx components/itemDetail.tsx
