+++ redux/counterSlice.js
@@ -1,0 +1,22 @@
+import { createSlice } from "@reduxjs/toolkit";
+
+export const counterSlice = createSlice({
+    name: "counter",
+    initialState: {
+        value: 1,
+    },
+    reducers: {
+        increment: (state) => {
+            state.value += 1;
+        },
+        decrement: (state) => {
+            state.value -= 1;
+        },
+        reset: (state) => {
+            state.value = 1;
+        }
+    }
+});
+
+export const { increment, decrement, reset } = counterSlice.actions;
+export default counterSlice.reducer;
