+++ redux/store.js
@@ -1,0 +1,10 @@
+import { configureStore } from "@reduxjs/toolkit";
+import chatSlice from "./chatSlice";
+import counterSlice from "./counterSlice";
+
+export const store = configureStore({
+    reducer: {
+        chatUserId: chatSlice,
+        counter: counterSlice
+    },
+});
